+++ changelog.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+* Don't switch to quoted-printable for long lines if already using base64
+* Fixed Travis-CI config when run on PHP 7
+* Add address parser for RFC822-format addresses
+* Update MS Office MIME types
+
+## Version 5.2.10 (May 4th 2015)
+* Add custom header getter
 * Use `application/javascript` for .js attachments
 * Improve RFC2821 compliance for timelimits, especially for end-of-data
 * Add Azerbaijani translations (Thanks to @mirjalal)
@@ -24,7 +31,14 @@
 * Improved checks and error messages for missing extensions
 * Store and report SMTP errors more consistently
 * Add MIME multipart preamble for better Outlook compatibility
+* Enable TLS encryption automatically if the server offers it
 * Provide detailed errors when individual recipients fail
+* Report more errors when connecting
+* Add extras classes to composer classmap
+* Expose stream_context_create options via new SMTPOptions property
+* Automatic encoding switch to quoted-printable if message lines are too long
+* Add Korean translation (Thanks to @ChalkPE)
+* Provide a pointer to troubleshooting docs on SMTP connection failure
 
 ## Version 5.2.9 (Sept 25th 2014)
 * **Important: The autoloader is no longer autoloaded by the PHPMailer class**